fix: use @types/git-raw-commits
